interface/session-manager-service/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.1
psycopg2-binary==2.9.6
psutil==5.9.5
requests==2.31.0